requirements.txt
certifi==2022.6.15.1
charset-normalizer==2.1.1
dateparser==1.1.1
idna==3.3
newscatcherapi==0.7.1
numpy==1.23.1
pandas==1.4.3
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
